+++ packages/pebble/src/tokenizer/utils/tokenIsAlsoIdentifier.ts
@@ -29,6 +29,7 @@ export function tokenIsAlsoIdentifier(token: Token): boolean {
         case Token.Withdraw:
         case Token.Propose:
         case Token.Vote:
+        case Token.Guard:
         // case Token.Context:
 
         // case Token.Type: // we don't allow it in pebble